move stun metrics
